+++ lib/components/Charts/AreaChart/index.tsx
